fix: names of tabs changed for more convinient ones and minor layout changes
