+++ frontend/src/features/admin/components/layout/AdminPostForm.tsx
@@ -1,0 +1,6 @@
+
+export function AdminPostForm(){
+    return(
+        <h1>FORMULARZ</h1>
+    )
+}
