+++ frontend/src/features/admin/pages/AdminAddPostPage.tsx
@@ -1,15 +1,17 @@
 import { FilePlus } from "lucide-react";
+import { AdminPostForm } from '../components/layout/AdminPostForm.tsx';
 
 export default function AdminAddPostPage() {
   return (
     <div className="space-y-6">
       <div className="flex items-center gap-3">
         <FilePlus className="h-6 w-6 text-muted-foreground" />
-        <h2 className="text-2xl font-semibold">Dodawanie wpisów</h2>
+        <h2 className="text-2xl font-semibold">Dodaj post</h2>
       </div>
       <p className="text-sm text-muted-foreground">
-        Tworzenie nowych wpisów i postów.
+        Tworzenie nowych postów.
       </p>
+      <AdminPostForm />
     </div>
   );
 }
